Rename the fees to product_fees.
